# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SkyArc Atlas API
  version: 1.0.0
  description: Internal DOOH Location Intelligence and Media Planning API
servers:
  - url: http://localhost:3001/api/v1
    description: Local development
tags:
  - name: Authentication
  - name: Users
  - name: Locations
  - name: LocationAssets
  - name: Screens
  - name: LocationIntelligence
paths:
  /health:
    get:
      tags: [Authentication]
      summary: Health check
      responses:
        "200":
          description: OK
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                deviceLabel: { type: string }
      responses:
        "200":
          description: Auth tokens
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      responses:
        "200":
          description: New access token
  /auth/logout:
    post:
      tags: [Authentication]
      summary: Logout
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Logged out
  /users:
    get:
      tags: [Users]
      summary: List users
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: User list
    post:
      tags: [Users]
      summary: Create user
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Created user
  /users/{id}:
    get:
      tags: [Users]
      summary: Get user
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: User
    patch:
      tags: [Users]
      summary: Update user
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Updated user
  /locations:
    get:
      tags: [Locations]
      summary: List locations
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Location list
    post:
      tags: [Locations]
      summary: Create or upsert location
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Location
  /locations/nearby:
    get:
      tags: [Locations]
      summary: Find nearby locations
      security: [{ bearerAuth: [] }]
      parameters:
        - name: lat
          in: query
          required: true
          schema: { type: number }
        - name: lng
          in: query
          required: true
          schema: { type: number }
        - name: radiusM
          in: query
          schema: { type: number, default: 1000 }
      responses:
        "200":
          description: Nearby locations
  /locations/{id}:
    get:
      tags: [Locations]
      summary: Get location
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Location
    patch:
      tags: [Locations]
      summary: Update location
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Updated location
  /locations/{id}/survey:
    put:
      tags: [Locations]
      summary: Upsert field survey
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Survey
    get:
      tags: [Locations]
      summary: Get survey
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Survey
  /locations/{id}/assets:
    get:
      tags: [LocationAssets]
      summary: List assets
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Asset list
  /locations/{id}/assets/presign:
    post:
      tags: [LocationAssets]
      summary: Request presigned upload URL
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Presigned URL
  /locations/{id}/assets/{assetId}/confirm:
    post:
      tags: [LocationAssets]
      summary: Confirm asset upload
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: assetId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Confirmed asset
  /locations/{id}/screens:
    get:
      tags: [Screens]
      summary: List screens
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Screen list
    post:
      tags: [Screens]
      summary: Create screen
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Screen
  /screens/{id}:
    patch:
      tags: [Screens]
      summary: Update screen
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Updated screen
  /screens/{id}/specification:
    put:
      tags: [Screens]
      summary: Upsert screen specification
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Specification
  /locations/{id}/attributes:
    get:
      tags: [LocationIntelligence]
      summary: List location attributes
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Attributes
  /locations/{id}/score:
    get:
      tags: [LocationIntelligence]
      summary: Get location score
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Score
  /locations/{id}/score/recompute:
    post:
      tags: [LocationIntelligence]
      summary: Recompute location score
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Score
  /locations/{id}/analyses:
    post:
      tags: [LocationIntelligence]
      summary: Queue AI analysis
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Analysis job
  /locations/{id}/analyses/{analysisId}:
    get:
      tags: [LocationIntelligence]
      summary: Get analysis status
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: analysisId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Analysis
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  path: { type: string }
                  message: { type: string }